H6 break fix
Removed a clear:both on H tags, as that was causing a paragraph break next to film trailers. I have no idea why a clear was put on a text tag, but if all of a sudden text looks wrong on the site, removing the clear might be the culprit.